Refactor client attributes to private
Change client attributes to private to encapsulate them within the `AcledClient` class. This helps ensure that the internal structure is hidden and maintained only through defined properties, improving code organization and maintainability.

The intent is to keep the unified interface with the API, keeping maintenance of each endpoint separate, and not forcing the user to have to see this more complicated underlying structure.